Refactors shipping order item dimensions to Integer
Updates the shipping order item dimensions (length, width,
height, weight) to use Integer instead of Double to align with
the GHN API requirements.

Sets the shipping order name to include the order ID for better
identification.

Rounds up double values to the nearest integer value.

+++ api/Shipping/src/main/java/com/lemoo/shipping/dto/common/ShippingOrderItem.java
@@ -16,8 +16,8 @@ public class ShippingOrderItem {
     private String name;
     private String code;
     private Integer quantity;
-    private Double length;
-    private Double width;
-    private Double weight;
-    private Double height;
+    private Integer length;
+    private Integer width;
+    private Integer weight;
+    private Integer height;
 }
